requirements.txt
opencv-python
python-dotenv
groq
langgraph
langchain-groq
pydub
SpeechRecognition
pyaudio
elevenlabs
gtts
gradio

-e .